Add Config file support
This is done with the good help from the Cereal project.
Cereal is a header-only serialisation tool that allows to dump/load the
contents of a class to/from a file.

Cereal is added as a submodule. run `git submodule init` to fetch cereal
from its remote repository.

Any new clone of this game project should be made with the `--recursive`
option set to automatically initialise git submodules.